Change return type and configure h2

+++ src/main/java/com/xpand/challenge/service/impl/ActorServiceImpl.java
@@ -4,9 +4,11 @@ import com.xpand.challenge.dto.ActorDTO;
 import com.xpand.challenge.model.Actor;
 import com.xpand.challenge.model.projection.ActorProjection;
 import com.xpand.challenge.repository.ActorRepository;
+import com.xpand.challenge.repository.MovieRepository;
 import com.xpand.challenge.service.ActorService;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -15,9 +17,11 @@ import java.util.stream.Collectors;
 public class ActorServiceImpl implements ActorService {
 
     private final ActorRepository actorRepository;
+    private final MovieRepository movieRepository;
 
-    public ActorServiceImpl(ActorRepository actorRepository) {
+    public ActorServiceImpl(ActorRepository actorRepository, MovieRepository movieRepository) {
         this.actorRepository = actorRepository;
+        this.movieRepository = movieRepository;
     }
 
     @Override
@@ -49,13 +53,17 @@ public class ActorServiceImpl implements ActorService {
         actorRepository.save(actor);
     }
 
+    @Transactional
     @Override
     public void deleteActor(Long id) {
-        try {
-            actorRepository.deleteById(id);
-        } catch (DataIntegrityViolationException e) {
-            throw new DataIntegrityViolationException("Invalid operation, actor linked to a movie");
+        boolean b = movieRepository
+                .findAll().stream()
+                .anyMatch(m -> m.getActors()
+                        .stream()
+                        .anyMatch(a -> a.getId().equals(id)));
+        if (b) {
+            throw new DataIntegrityViolationException("Unable to delete an anime linked to a movie");
         }
-
+        actorRepository.deleteById(id);
     }
 }
